requirements.txt
tensorflow-gpu==2.2.0
keras
numpy
opencv-contrib-python
scikit-learn
matplotlib